+++ packages/engine/src/index.ts
@@ -63,7 +63,7 @@ export * from './modules/index.js';
 export * from './modules/core-econ.js';
 export * from './modules/core-pending.js';
 export * from './modules/core-events.js';
-export { loansOf } from './modules/loans.js';
+export { defaultDebtOf, loansOf, totalOwed } from './modules/loans.js';
 export { weeklySubTotal } from './modules/subscriptions.js';
 export { wellbeingBand, wellbeingOf } from './modules/wellbeing.js';
 export { coreSetup } from './modules/core-setup.js';